Single relay for range and values specimen builders
